% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrar_tablas_patron.R
\name{filtrar_tabla}
\alias{filtrar_tabla}
\title{Filtrar las tablas por emi, sin y datos generales}
\usage{
filtrar_tabla(tabla, pattern)
}
\arguments{
\item{pattern}{}
}
\description{
Filtrar las tablas por emi, sin y datos generales
}
