% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_variadas.R
\name{graph_error}
\alias{graph_error}
\title{GGplot formato erroneo}
\usage{
graph_error(data, titles)
}
\arguments{
\item{data}{tabla con el numero de columnas correctas e incorrectas}

\item{titles}{ramo al que pertenece}
}
\description{
GGplot formato erroneo
}
