% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_variadas.R
\name{list_names}
\alias{list_names}
\title{Poner el nombre del ramo a cada elemento de la lista}
\usage{
list_names(data)
}
\arguments{
\item{data}{}
}
\description{
Poner el nombre del ramo a cada elemento de la lista
}
