% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_read_dge.R
\name{mod_dge_read_files_server}
\alias{mod_dge_read_files_server}
\title{Modulo para leer las tablas de Datos generales}
\usage{
mod_dge_read_files_server(id, dge_table)
}
\arguments{
\item{dge_table}{}
}
\value{

}
\description{
Modulo para leer las tablas de Datos generales
}
